Introduce JLargeArrayInstruction instruction
