+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/model/BIRVisitor.java
@@ -169,10 +169,6 @@ public abstract class BIRVisitor {
         throw new AssertionError();
     }
 
-    public void visit(BIRNonTerminator.NewLargeArray birNewLargeArray) {
-        throw new AssertionError();
-    }
-
     public void visit(BIRNonTerminator.FieldAccess birFieldAccess) {
         throw new AssertionError();
     }
