+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/model/InstructionKind.java
@@ -123,8 +123,6 @@ public enum InstructionKind {
     NEW_RE_FLAG_ON_OFF((byte) 100),
     NEW_RE_QUANTIFIER((byte) 101),
 
-    NEW_LARGE_ARRAY((byte) 102),
-
     PLATFORM((byte) 128);
 
     byte value;
